+++ src/jotai/atom.ts
@@ -25,6 +25,6 @@ export const recitersSortAtom = createAtomWithStorage<
   'popular' | 'alphabetical' | 'views'
 >('reciters-sort-atom', 'popular');
 export const selectedReciterAtom = createAtomWithStorage<Reciter | null>(
-  'selected-riwaya',
+  'selected-reciter',
   null
 );